Fix methanol benchmark summary import (#70)
